Fixed the issues when building in Msys2 or Cygwin.
Supported the MinGW64, MINGW32, UCRT64, MSYS and CYGWIN.
Tested in Msys2/3.4.10-2, Cygwin 3.3.6/Install 2.919 in Windowns 7.
Fixed the dependency of iupwin_image_wdl.c in config.mak.
Fixed the pointer conversion in 'iupwin_image.c' and 'iupwin_info.c'
Workarounded the problem for Cygwin and Msys that tOleHandler.cpp could
not find the definition of byte/boolean and 'hyper'